+++ CHANGELOG.md
@@ -14,6 +14,8 @@ Possible sections are:
 <!-- next-header -->
 ## [Unreleased] - ReleaseDate
 
+## [0.3.0] - 2023-01-31
+
 ### Added
 
 - `Resolution::area_rads2`